Move shared validation for StridedSliceOp to a separate header, in preparation for calling it from the shape fn. Change: 132732667
